+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Features:
   - Added `RxDatabase.remove()` [docs](https://pubkey.github.io/rxdb/RxDatabase.html#remove)
   - Added options for `RxCollection.sync()`: direction and pouchdb-replication-settings [docs](https://pubkey.github.io/rxdb/RxCollection.html#sync)
   - Added query-param for `RxCollection.sync()` to allow selector-based replication [docs](https://pubkey.github.io/rxdb/RxCollection.html#sync)
+  - Added TemporaryDocuments `RxCollection.newDocument()` [docs](https://pubkey.github.io/rxdb/RxCollection.html#newDocument)
 
 Other:
   - Added .babelrc to .npmignore